% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cite_packages.R
\name{cite_pkg}
\alias{cite_pkg}
\title{Cite packages in .docx output}
\usage{
cite_pkg(pkg_list)
}
\arguments{
\item{pkg_list}{List of used packages}
}
\value{
None
}
\description{
Cite packages in .docx output
}
\examples{
pkg <- c("vegan", "ggplot2")
cite_pkg(pkg)
}
