% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_csl.R
\name{download_csl}
\alias{download_csl}
\title{Download csl citation style}
\usage{
download_csl(journal_style, directory)
}
\arguments{
\item{journal_style}{Journal style (see examples)}

\item{directory}{Directory to store the csl file}
}
\value{
None
}
\description{
Download csl citation style from \url{https://github.com/citation-style-language/styles}.
}
\examples{
# First, search by our prefered citation style using:
eco <- grep(pattern = "ecology", x = rcrossref::get_styles(), value = TRUE)
# then:
download_csl(journal_style = eco[4], directory = "")
}
