% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_comments.R
\name{extract_comments}
\alias{extract_comments}
\title{Extract comments of a docx file to a csv}
\usage{
extract_comments(docx_path, filename_to_save = NULL)
}
\arguments{
\item{docx_path}{docx path}

\item{filename_to_save}{name of file to save}
}
\value{
a csv file
}
\description{
Extract comments of a docx file to a csv
}
