% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_comments.R
\name{extract_comments_from_docx}
\alias{extract_comments_from_docx}
\title{Extract comments of a docx file to a csv}
\usage{
extract_comments_from_docx(path)
}
\arguments{
\item{path}{a character vector with path to docx files}
}
\value{
a csv file
}
\description{
Extract comments of a docx file to a csv
}
