% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_template.R
\name{setup_template}
\alias{setup_template}
\title{Setup a new project template}
\usage{
setup_template(name)
}
\arguments{
\item{name}{Name of the projetc}
}
\value{
None
}
\description{
Setup and open a new projetc in RStudio
}
\examples{
\dontrun{
setup_template("my_new_project")
}
}
