% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_wdiff.R
\name{setup_wdiff}
\alias{setup_wdiff}
\title{Setup wdiff for .docx files via pre- and post-commit hooks}
\usage{
setup_wdiff(verbose = TRUE)
}
\arguments{
\item{verbose}{If TRUE, (default) show global and local git config files}
}
\value{
None
}
\description{
Setup wdiff for .docx files via pre- and post-commit hooks
}
\examples{
# Just run:
setup_wdiff()
}
