% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipak.R
\name{ipak}
\alias{ipak}
\title{Install and load multiple R packages}
\usage{
ipak(pkg)
}
\arguments{
\item{pkg}{A character vector with the package names}
}
\value{
None
}
\description{
Thanks to Steve Worthington \url{https://gist.github.com/stevenworthington/3178163}.
}
\examples{
\dontrun{
list_pkg <- c("vegan", "ggplot2")
ipak(list_pkg)
}
}
