% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_template.R
\name{setup_template}
\alias{setup_template}
\title{Setup a new project template}
\usage{
setup_template(name, open = TRUE)
}
\arguments{
\item{name}{Name of the projetc}

\item{open}{If the new must be open in a new RStudio session}
}
\value{
None
}
\description{
Setup and open a new projetc in RStudio
}
\examples{
\dontrun{
setup_template(name = "my_new_project", open = FALSE)
}
}
